+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/darktraffic/main/DarkServerTestModule.scala
@@ -1,10 +1,11 @@
 package com.twitter.finatra.http.tests.integration.darktraffic.main
 
 import com.twitter.conversions.time._
+import com.twitter.finagle.Http
 import com.twitter.finagle.http.Method.{Delete, Post}
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.http.modules.DarkTrafficFilterModule
-import com.twitter.util.Duration
+import com.twitter.inject.Injector
 
 object DarkServerTestModule extends DarkTrafficFilterModule {
 
@@ -12,12 +13,27 @@ object DarkServerTestModule extends DarkTrafficFilterModule {
    * Function to determine if the request should be "sampled", e.g.
    * sent to the dark service.
    */
-  override val enableSampling: Request => Boolean = { request =>
+  override def enableSampling(injector: Injector): Request => Boolean = { request =>
     request.method match {
       case Post | Delete => false
       case _ => true
     }
   }
-  override val acquisitionTimeout: Duration = 100.millis
-  override val requestTimeout: Duration = 100.millis
+
+  /**
+   * Override to specify further configuration of the Finagle [[Http.Client]].
+   *
+   * @param injector the [[com.twitter.inject.Injector]] for use in configuring the underlying client.
+   * @param client   the default configured [[Http.Client]].
+   *
+   * @return a configured instance of the [[Http.Client]]
+   */
+  override protected def configureHttpClient(
+    injector: Injector,
+    client: Http.Client
+  ): Http.Client = {
+    client
+      .withSession.acquisitionTimeout(100.millis)
+      .withRequestTimeout(100.millis)
+  }
 }